fix all comments from git
1. Remove testfixtures
2. In render test group test in one and use TestCase
3. Move some methods from md to tags
4. Move Parser to new file
