+++ cs/Markdown/MarksownParserTests/ReadAllBoldTokensTest.cs
@@ -3,7 +3,6 @@ using NUnit.Framework;
 
 namespace Markdown
 {
-    [TestFixture]
     public class ReadAllBoldTokensTest
     {
         [Test]
