+++ cs/Markdown/MarksownParserTests/ReadAllItalicTokensTest.cs
@@ -3,7 +3,6 @@ using NUnit.Framework;
 
 namespace Markdown
 {
-    [TestFixture]
     public class ReadAllItalicTokensTest
     {
         [Test]
