+++ cs/Markdown/MarksownParserTests/ReadHeaderTokenTest.cs
@@ -4,7 +4,6 @@ using NUnit.Framework;
 
 namespace Markdown
 {
-    [TestFixture]
     public class ReadHeaderToken
     {
         [Test]
@@ -19,7 +18,7 @@ namespace Markdown
             var result = MarkdownParser.ReadHeaderToken("#asdf");
             result.Length.Should().Be(2);
             result[0].Should().BeEquivalentTo(new Tag("<h1>", 0, 1));
-            result[1].Should().BeEquivalentTo(new Tag("</h1>", 5, 0));
+            result[1].Should().BeEquivalentTo(new Tag("</h1>", 5, 1));
         }
         
     }
